Fix race condition in RemoteTransaction. Use agrpc::run in execute_loop_single_threaded
